requirements.txt
streamlit
chromadb
requests
google-generativeai
bs4
